requirements.txt
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
dotenv==0.9.9
exceptiongroup==1.2.2
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
idna==3.10
Jinja2==3.1.6
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
outcome==1.3.0.post0
pandas==2.2.3
pillow==11.1.0
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
